Nyla/requirements.txt
# MySQL connector for Python
mysql-connector-python

# Data manipulation library
pandas
